Some cleanup and new GetPropertyValue-to-long-value-LookupHash
GetPropertyValue-to-long-value-LookupHash creates a hash
which is configurable. By default is a hash of long propnames
as the keys. below that level, there are property value aliases, which
are the keys and then map to the long property alias names.

Bidi_Class => ${:AL("Arabic_Letter"), :AN("Arabic_Number"),
:Arabic_Letter("Arabic_Letter"), :Arabic_Number("Arabic_Number"),
:B("Paragraph_Separator"), :BN("Boundary_Neutral") }

You can also configure to use short property names instead of long.
Such as bc instead of Bidi_Class as above.

You can also configure it to have the alias property value names map
to the shorthand names. So "Arabic_Letter" => "AL" instead of the reverse
as shown above.